minor readability improvements for blueprints
